edge_ingestion_service/requirements.txt
flask
flask-cors
paho-mqtt<2.0
psutil
requests
influxdb-client